pyproject.toml
[tool.poetry]
name = "firehol-service"
version = "0.1.0"
description = "Service to get data from firehol"
authors = ["Stephen O'Malley <stephenomalley81@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "^2.133.0"
aws-cdk-aws-lambda-python-alpha = "^2.133.0a0"
gitpython = "^3.1.42"
pydantic = "^2.6.4"
boto3 = {extras = ["crt"], version = "^1.34.68"}
boto3-stubs = {extras = ["dynamodb", "events", "s3"], version = "^1.34.68"}


[tool.poetry.group.lint.dependencies]
ruff = "^0.3.3"
boto3-stubs = {extras = ["dynamodb", "s3"], version = "^1.34.68"}

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"


[tool.poetry.group.dev.dependencies]
aws-lambda-powertools = "^2.35.1"
pre-commit = "^3.6.2"
mypy = "^1.9.0"
boto3 = {extras = ["s3"], version = "^1.34.68"}


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.8
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I", "Q"]
ignore = ["F821"] # This is temp

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
force-single-line = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
